use Lua stdlib instead of Neovim API for request parsing speed-up
Specifically, use `string.gmatch` instead of `ipairs(vim.split(...))`.

In a crude microbenchmark, `process_request_line` takes about 70 percent
less time to run with the former than with the latter.  As to the
performance of the web server, measured with `siege -c 100`, the
transaction rate is also higher, above 8,000 transactions/second for the
first 100,000 or so transactions.